pyproject.toml
[tool.poetry]
name = "agent-test"
version = "0.1.0"
description = "Production-ready RAG agent with FastAPI, FAISS, and Twilio"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "agent_test", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# Web backend
fastapi = "^0.111.0"
uvicorn = "^0.29.0"

# LangChain + RAG
langchain = "^0.2.0"
langchain-community = "^0.2.0"
openai = "^1.30.1"
tiktoken = "^0.7.0"

# Vector store
faiss-cpu = "^1.8.0"

# Twilio
twilio = "^9.0.4"

# Database
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"

# Templates & Static Files
jinja2 = "^3.1.4"
aiofiles = "^23.2.0"

# Authentication & Security
passlib = "^1.7.4"
bcrypt = "^4.1.0"
python-jose = "^3.3.0"

# Env & Logging
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
